Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/127-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected end of script

Harness Error (FAIL), message = SyntaxError: Unexpected end of script

FAIL scheduler: appending non-text children to script  assert_array_equals: lengths differ, expected array ["inline script #1", "inline script #2", "end inline script #2", "end inline script #1"] length 4, got ["inline script #1", "end inline script #1"] length 2
